Comment out test.types.ts @mourner
