+++ .gitignore
@@ -1,4 +1,3 @@
 node_modules
 npm-debug.log
-test.types.js
 yarn.lock
